fix(warnings): eliminate all compiler and linker warnings across 28 files
- boot.asm / isr.asm / context.asm: add .note.GNU-stack section to suppress
  linker "executable stack" warnings; declare .boot.bss with `nobits` flag
- include/fs/vfs.h: wrap O_* flag defines in #ifndef guards to stop redefinition
  warnings when syscall.h is included first
- net/ip.c, net/socket.c, net/dhcp.c, net/dns.c: add missing #include <net/udp.h>
  to resolve implicit-declaration warnings for udp_send/udp_receive/udp_recv_poll
- net/tcp.c: remove unnecessary __attribute__((packed)) from naturally-aligned
  pseudo-header struct to fix -Waddress-of-packed-member
- apps/clock.c: trim sin_tab[360] initializer from 368 to 360 elements
- apps/terminal.c, apps/filemanager.c, apps/texteditor.c: change keycode
  variable from uint8_t to int so KEY_UP/DOWN_ARROW (0x100/0x101) comparisons
  are not always-false due to truncation
- apps/calculator.c: drop redundant `bid >= BID_0` (always true for uint8_t)
- kernel/kernel.c: add (uint64_t) cast on PHYS_TO_VIRT assignment; fix \_
  unknown escape sequence in ASCII banner string
- kernel/log.c: remove unused level_colors array
- interrupts/pic.c: suppress unused mask1/mask2 with (void) casts
- drivers/ata.c: suppress unused bus_idx param with (void) cast
- drivers/pci.c: mark pci_read16 with __attribute__((unused))
- process/process.c: suppress unused argv param with (void) cast
- scheduler/scheduler.c: suppress unused regs param with (void) cast
- apps/sysmonitor.c: suppress unused s param with (void) cast
- gui/desktop.c: remove unused g_first_frame static variable
- gui/notify.c: actually use body_col in draw_string call (was dead assignment)
- services/splash.c: remove unused start_ticks variable
- libc/printf.c: cast loop bound to size_t in PUTS macro to fix -Wsign-compare
- net/socket.c: remove always-false `next_port > 65535` check for uint16_t
- fs/fat32.c: remove unused lfn_pos; rename inner ci to dir_ci to fix -Wshadow;
  mark fat32_resolve/lfn_collect_chars with __attribute__((unused))
- fs/ramfs.c: remove /* inside block comments to fix -Wcomment

Build result: 0 errors, 0 warnings. kernel.elf (351K) + aetheros.iso (720K).

+++ apps/filemanager.c
@@ -411,8 +411,8 @@ static void fm_on_event(wid_t wid, gui_event_t* evt, void* ud)
     }
 
     case GUI_EVENT_KEY_DOWN: {
-        uint8_t kc = evt->key.keycode;
-        char ch    = evt->key.ch;
+        int kc  = evt->key.keycode;
+        char ch = evt->key.ch;
 
         if (kc == KEY_UP_ARROW) {
             if (f->selected > 0) f->selected--;
