+++ apps/terminal.c
@@ -654,8 +654,8 @@ static void term_on_event(wid_t wid, gui_event_t* evt, void* ud)
         break;
 
     case GUI_EVENT_KEY_DOWN: {
-        uint8_t kc = evt->key.keycode;
-        char ch    = evt->key.ch;
+        int kc  = evt->key.keycode;
+        char ch = evt->key.ch;
 
         if (kc == KEY_UP_ARROW) {
             /* History navigation */
