+++ apps/texteditor.c
@@ -154,7 +154,7 @@ static void te_on_event(wid_t wid, gui_event_t* evt, void* ud)
         break;
 
     case GUI_EVENT_KEY_DOWN: {
-        uint8_t kc = evt->key.keycode;
+        int kc  = evt->key.keycode;
         char ch = evt->key.ch;
 
         if (kc == KEY_UP_ARROW) {
